Crate dyn_utils

A utility library for working with trait objects.

Trait objects, i.e. dyn Trait, are unsized and requires to be stored in a container like Box. This crate provides DynObject, a container for trait object with a generic storage.

storage::Raw stores object in place, making DynObject<dyn Trait, storage::Raw> allocation-free. On the other hand, storage::RawOrBox falls back to an allocated Box if the object is too big to fit in place.

Saving one allocation makes DynObject a good alternative to Box when it comes to write a dyn-compatible version of a trait with return-position impl Trait, such as async methods.

Examples

use dyn_utils::DynObject;

trait Callback {
    fn call(&self, arg: &str) -> impl Future<Output = ()> + Send;
}

// Dyn-compatible version
trait DynCallback {
    fn call<'a>(&'a self, arg: &'a str) -> DynObject<dyn Future<Output = ()> + Send + 'a>;
}

impl<T: Callback> DynCallback for T {
    fn call<'a>(&'a self, arg: &'a str) -> DynObject<dyn Future<Output = ()> + Send + 'a> {
        DynObject::new(self.call(arg))
    }
}

async fn exec_callback(callback: &dyn DynCallback) {
    callback.call("Hello world!").await;
}

This crate also provides dyn_trait proc-macro to do the same as above:

#[dyn_utils::dyn_trait] // generates `DynCallback` trait
trait Callback {
    fn call(&self, arg: &str) -> impl Future<Output = ()> + Send;
}

async fn exec_callback(callback: &dyn DynCallback) {
    callback.call("Hello world!").await;
}

Modules

storage

The storages backing DynObject.

Structs

DynObject

A trait object whose data is stored in a generic Storage.

Attribute Macros

dyn_object

Make a trait compatible with DynObject.

dyn_trait

Generate a dyn compatible trait from a given trait declaration.

sync

Mark an async method as internally synchronous.