dyn_utils::object

Trait Vtable

pub unsafe trait Vtable<T>: DynTrait {
    const DROP_IN_PLACE_FN: Option<unsafe fn(NonNull<()>)> = _;

    // Required method
    fn vtable<S: Storage>() -> &'static Self::Vtable;
}

A vtable constructor.

Safety

DynTrait::layout must return core::alloc::Layout::new::<T>().

DynTrait::drop_in_place_fn must returns <Self as crate::object::Vtable<T>>::DROP_IN_PLACE_FN

Provided Associated Constants

const DROP_IN_PLACE_FN: Option<unsafe fn(NonNull<()>)> = _

The function has the same safety contract that core::ptr::drop_in_place.

Required Methods

fn vtable<S: Storage>() -> &'static Self::Vtable

Returns the vtable for a given T stored in S.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.