dyn_utils

Module storage

The storages backing DynObject.

Structs

Align

A zero-sized-type aligned to N. Compound types containing a field Align<N> with have an alignment of at least N.

Boxalloc

A type-erased Box.

Raw

A raw storage, where data is stored in place.

RawOrBox

A Raw storage with Box backup if the object doesn’t fit in.

Traits

Alignment

Implemented for all Align<N> where N is a valid alignment (i.e., a power of two less-than-or-equal to 228).

FromBoxalloc

A storage that can be constructed from boxed data.

Storage

A storage that can be used to store dynamic type-erased objects.

Type Aliases

DefaultStorage

Default storage for DynObject, and used in dyn_trait macro.