dyn_utils::storage

Trait FromBox

pub trait FromBox: Storage {
    // Required method
    fn from_box<T>(boxed: StdBox<T>) -> Self;
}

Available on crate feature alloc only.

A storage that can be constructed from boxed data.

Required Methods

fn from_box<T>(boxed: StdBox<T>) -> Self

Constructs a new storage storing T.

Data may be moved out the box if it fits in the storage.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.